Add ArrayableRule
