+++ src/Illuminate/Validation/Rules/In.php
@@ -2,61 +2,20 @@
 
 namespace Illuminate\Validation\Rules;
 
-use BackedEnum;
-use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Traits\Conditionable;
 use Stringable;
-use UnitEnum;
 
 class In implements Stringable
 {
-    /**
-     * The name of the rule.
-     *
-     * @var string
-     */
-    protected $rule = 'in';
-
-    /**
-     * The accepted values.
-     *
-     * @var array
-     */
-    protected $values;
-
-    /**
-     * Create a new in rule instance.
-     *
-     * @param  \Illuminate\Contracts\Support\Arrayable|\BackedEnum|\UnitEnum|array|string  $values
-     * @return void
-     */
-    public function __construct($values)
-    {
-        if ($values instanceof Arrayable) {
-            $values = $values->toArray();
-        }
-
-        $this->values = is_array($values) ? $values : func_get_args();
-    }
+    use ArrayableRule, Conditionable;
 
     /**
      * Convert the rule to a validation string.
      *
      * @return string
-     *
-     * @see \Illuminate\Validation\ValidationRuleParser::parseParameters
      */
     public function __toString()
     {
-        $values = array_map(function ($value) {
-            $value = match (true) {
-                $value instanceof BackedEnum => $value->value,
-                $value instanceof UnitEnum => $value->name,
-                default => $value,
-            };
-
-            return '"'.str_replace('"', '""', $value).'"';
-        }, $this->values);
-
-        return $this->rule.':'.implode(',', $values);
+        return 'in:' . implode(',', $this->formatValues());
     }
 }
