+++ src/Illuminate/Validation/Rules/NotIn.php
@@ -2,41 +2,12 @@
 
 namespace Illuminate\Validation\Rules;
 
-use BackedEnum;
-use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Traits\Conditionable;
 use Stringable;
-use UnitEnum;
 
 class NotIn implements Stringable
 {
-    /**
-     * The name of the rule.
-     *
-     * @var string
-     */
-    protected $rule = 'not_in';
-
-    /**
-     * The accepted values.
-     *
-     * @var array
-     */
-    protected $values;
-
-    /**
-     * Create a new "not in" rule instance.
-     *
-     * @param  \Illuminate\Contracts\Support\Arrayable|\BackedEnum|\UnitEnum|array|string  $values
-     * @return void
-     */
-    public function __construct($values)
-    {
-        if ($values instanceof Arrayable) {
-            $values = $values->toArray();
-        }
-
-        $this->values = is_array($values) ? $values : func_get_args();
-    }
+    use ArrayableRule, Conditionable;
 
     /**
      * Convert the rule to a validation string.
@@ -45,16 +16,6 @@ class NotIn implements Stringable
      */
     public function __toString()
     {
-        $values = array_map(function ($value) {
-            $value = match (true) {
-                $value instanceof BackedEnum => $value->value,
-                $value instanceof UnitEnum => $value->name,
-                default => $value,
-            };
-
-            return '"'.str_replace('"', '""', $value).'"';
-        }, $this->values);
-
-        return $this->rule.':'.implode(',', $values);
+        return 'not_in:'.implode(',', $this->formatValues());
     }
 }
